Implemented abstract undo/redo class
